address kurt's comments
