+++ flink-table/flink-sql-client/src/main/java/org/apache/flink/table/client/cli/CliTableResultView.java
@@ -22,6 +22,7 @@ import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.table.client.gateway.ResultDescriptor;
 import org.apache.flink.table.client.gateway.SqlExecutionException;
 import org.apache.flink.table.client.gateway.TypedResult;
+import org.apache.flink.table.utils.PrintUtils;
 import org.apache.flink.types.Row;
 
 import org.jline.keymap.KeyMap;
@@ -262,7 +263,7 @@ public class CliTableResultView extends CliResultView<CliTableResultView.ResultT
 		// convert page
 		final List<String[]> stringRows = rows
 				.stream()
-				.map(CliUtils::rowToString)
+				.map(PrintUtils::rowToString)
 				.collect(Collectors.toList());
 
 		// update results
