+++ flink-table/flink-sql-client/src/main/resources/META-INF/NOTICE
@@ -9,4 +9,3 @@ See bundled license files for details.
 
 - org.jline:jline-terminal:3.9.0
 - org.jline:jline-reader:3.9.0
-- com.ibm.icu:icu4j:65.1
