+++ flink-table/flink-table-api-java/pom.xml
@@ -43,13 +43,6 @@ under the License.
 			<version>${project.version}</version>
 		</dependency>
 
-		<!-- Tools to unify display format for different languages -->
-		<dependency>
-			<groupId>com.ibm.icu</groupId>
-			<artifactId>icu4j</artifactId>
-			<version>65.1</version>
-		</dependency>
-
 		<!-- test dependencies -->
 
 		<dependency>
