+++ flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/api/internal/TableResultImpl.java
@@ -22,20 +22,15 @@ import org.apache.flink.annotation.Internal;
 import org.apache.flink.core.execution.JobClient;
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.api.ResultKind;
-import org.apache.flink.table.api.TableColumn;
 import org.apache.flink.table.api.TableResult;
 import org.apache.flink.table.api.TableSchema;
-import org.apache.flink.table.utils.EncodingUtils;
+import org.apache.flink.table.utils.PrintUtils;
 import org.apache.flink.types.Row;
 import org.apache.flink.util.Preconditions;
 
-import com.ibm.icu.lang.UCharacter;
-import com.ibm.icu.lang.UProperty;
-import org.apache.commons.lang3.StringUtils;
-
 import javax.annotation.Nullable;
 
-import java.util.ArrayList;
+import java.io.PrintWriter;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
@@ -51,11 +46,6 @@ public class TableResultImpl implements TableResult {
 			.tableSchema(TableSchema.builder().field("result", DataTypes.STRING()).build())
 			.data(Collections.singletonList(Row.of("OK")))
 			.build();
-
-	// constants for printing
-	private static final String NULL_COLUMN = "(NULL)";
-	private static final int MAX_COLUMN_WIDTH = 30;
-	private static final String COLUMN_TRUNCATED_FLAG = "...";
 
 	private final JobClient jobClient;
 	private final TableSchema tableSchema;
@@ -96,7 +86,7 @@ public class TableResultImpl implements TableResult {
 	@Override
 	public void print() {
 		Iterator<Row> it = collect();
-		doPrint(it);
+		PrintUtils.printAsTableauForm(getTableSchema(), it, new PrintWriter(System.out));
 	}
 
 	public static Builder builder() {
@@ -177,158 +167,4 @@ public class TableResultImpl implements TableResult {
 		}
 	}
 
-	/**
-	 * Displays the result in a tableau form.
-	 * For example:
-	 * +-------------+---------+-------------+
-	 * | boolean_col | int_col | varchar_col |
-	 * +-------------+---------+-------------+
-	 * |        true |       1 |         abc |
-	 * |       false |       2 |         def |
-	 * |      (NULL) |  (NULL) |      (NULL) |
-	 * +-------------+---------+-------------+
-	 * 3 row(s) in result
-	 *
-	 * <p>Changelog is not supported until FLINK-16998 is finished.
-	 */
-	private void doPrint(Iterator<Row> it) {
-		List<String[]> rows = new ArrayList<>();
-
-		// fill field names first
-		List<TableColumn> columns = getTableSchema().getTableColumns();
-		rows.add(columns.stream().map(TableColumn::getName).toArray(String[]::new));
-		while (it.hasNext()) {
-			rows.add(rowToString(it.next()));
-		}
-
-		int[] colWidths = columnWidthsByContent(columns, rows);
-		String borderline = genBorderLine(colWidths);
-
-		// print field names
-		System.out.println(borderline);
-		printSingleRow(colWidths, rows.get(0));
-		System.out.println(borderline);
-
-		// print content
-		if (rows.size() > 1) {
-			rows.subList(1, rows.size()).forEach(row -> printSingleRow(colWidths, row));
-			System.out.println(borderline);
-		}
-
-		System.out.println((rows.size() - 1) + " row(s) in result");
-	}
-
-	private String[] rowToString(Row row) {
-		final String[] fields = new String[row.getArity()];
-		for (int i = 0; i < row.getArity(); i++) {
-			final Object field = row.getField(i);
-			if (field == null) {
-				fields[i] = NULL_COLUMN;
-			} else {
-				fields[i] = EncodingUtils.objectToString(field);
-			}
-		}
-		return fields;
-	}
-
-	private int[] columnWidthsByContent(List<TableColumn> columns, List<String[]> rows) {
-		// fill width with field names first
-		int[] colWidths  = columns.stream().mapToInt(col -> col.getName().length()).toArray();
-
-		// fill column width with real data
-		for (String[] row : rows) {
-			for (int i = 0; i < row.length; ++i) {
-				colWidths[i] = Math.max(colWidths[i], getStringDisplayWidth(row[i]));
-			}
-		}
-
-		// adjust column width with maximum length
-		for (int i = 0; i < colWidths.length; ++i) {
-			colWidths[i] = Math.min(colWidths[i], MAX_COLUMN_WIDTH);
-		}
-
-		return colWidths;
-	}
-
-	private String genBorderLine(int[] colWidths) {
-		StringBuilder sb = new StringBuilder();
-		sb.append("+");
-		for (int width : colWidths) {
-			sb.append(StringUtils.repeat('-', width + 1));
-			sb.append("-+");
-		}
-		return sb.toString();
-	}
-
-	private void printSingleRow(int[] colWidths, String[] cols) {
-		StringBuilder sb = new StringBuilder();
-		sb.append("|");
-		int idx = 0;
-		for (String col : cols) {
-			sb.append(" ");
-			int displayWidth = getStringDisplayWidth(col);
-			if (displayWidth <= colWidths[idx]) {
-				sb.append(StringUtils.repeat(' ', colWidths[idx] - displayWidth));
-				sb.append(col);
-			} else {
-				sb.append(truncateString(col, colWidths[idx] - COLUMN_TRUNCATED_FLAG.length())); //
-				sb.append(COLUMN_TRUNCATED_FLAG);
-			}
-			sb.append(" |");
-			idx++;
-		}
-		System.out.println(sb.toString());
-	}
-
-	private String truncateString(String col, int targetWidth) {
-		int passedWidth = 0;
-		int i = 0;
-		for (; i < col.length(); i++) {
-			if (isFullWidth(Character.codePointAt(col, i))) {
-				passedWidth += 2;
-			} else {
-				passedWidth += 1;
-			}
-			if (passedWidth >= targetWidth) {
-				break;
-			}
-		}
-		String substring = col.substring(0, i);
-
-		// pad with ' ' before the column
-		int lackedWidth = targetWidth - getStringDisplayWidth(substring);
-		if (lackedWidth > 0){
-			substring = StringUtils.repeat(' ', lackedWidth) + substring;
-		}
-		return substring;
-	}
-
-	private int getStringDisplayWidth(String str) {
-		int numOfFullWidthCh = (int) str.codePoints().filter(this::isFullWidth).count();
-		return str.length() + numOfFullWidthCh;
-	}
-
-	/**
-	 * Check codePoint is FullWidth or not according to Unicode Standard version 12.0.0.
-	 * See http://unicode.org/reports/tr11/
-	 */
-	private boolean isFullWidth(int codePoint) {
-		int value = UCharacter.getIntPropertyValue(codePoint, UProperty.EAST_ASIAN_WIDTH);
-		switch (value) {
-			case UCharacter.EastAsianWidth.NEUTRAL:
-				return false;
-			case UCharacter.EastAsianWidth.AMBIGUOUS:
-				return false;
-			case UCharacter.EastAsianWidth.HALFWIDTH:
-				return false;
-			case UCharacter.EastAsianWidth.FULLWIDTH:
-				return true;
-			case UCharacter.EastAsianWidth.NARROW:
-				return false;
-			case UCharacter.EastAsianWidth.WIDE:
-				return true;
-			default:
-				throw new RuntimeException("unknown UProperty.EAST_ASIAN_WIDTH: " + value);
-		}
-	}
 }
