+++ flink-table/flink-table-common/pom.xml
@@ -51,6 +51,15 @@ under the License.
 			<!-- managed version -->
 		</dependency>
 
+		<!-- Tools to unify display format for different languages -->
+		<dependency>
+			<groupId>com.ibm.icu</groupId>
+			<artifactId>icu4j</artifactId>
+			<version>65.1</version>
+			<!-- set provided here, and will package it into planner jar -->
+			<scope>provided</scope>
+		</dependency>
+
 		<dependency>
 			<groupId>org.apache.flink</groupId>
 			<artifactId>flink-core</artifactId>
