+++ flink-table/flink-table-common/src/main/resources/META-INF/NOTICE
