+++ flink-table/flink-table-common/src/main/resources/META-INF/licenses/LICENSE.icu4j
