+++ flink-table/flink-table-uber/pom.xml
@@ -105,6 +105,13 @@ under the License.
 									<include>org.apache.flink:flink-cep_${scala.binary.version}</include>
 								</includes>
 							</artifactSet>
+							<relocations>
+								<relocation>
+									<!-- icu4j's dependencies -->
+									<pattern>com.ibm.icu</pattern>
+									<shadedPattern>org.apache.flink.table.shaded.com.ibm.icu</shadedPattern>
+								</relocation>
+							</relocations>
 						</configuration>
 					</execution>
 				</executions>
